Distinguish ChatException and ChatError
Jackson keeps analyzing the whole Exception object, which just takes too long.
