+++ src/main/java/com/penguineering/hareairis/model/ChatError.java
@@ -6,41 +6,19 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+@AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ChatError extends RuntimeException {
-    @Getter
-    @AllArgsConstructor
-    public enum Code {
-        CODE_BAD_REQUEST(400),
-        CODE_TOO_MANY_REQUESTS(429),
-        CODE_INTERNAL_SERVER_ERROR(500),
-        CODE_GATEWAY_TIMEOUT(504);
-
-        private final int code;
-
-    }
-
+public class ChatError {
     @JsonProperty("code")
     private int code;
 
     @JsonProperty("message")
     private String message;
 
-    public ChatError(String message) {
-        super(message);
-        this.code = Code.CODE_INTERNAL_SERVER_ERROR.getCode();
-        this.message = message;
-    }
-
-    public ChatError(Code code, String message) {
-        super(message);
-        this.code = code.getCode();
-        this.message = message;
-    }
-
-    public boolean is5xxServerError() {
-        return code >= 500 && code < 600;
+    public ChatError(ChatException ex) {
+        this.code = ex.getCode();
+        this.message = ex.getMessage();
     }
 }
